Lab2/requirements.txt
absl-py==2.1.0
pytest==8.3.3
python-docx==1.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
httpx==0.24.1
openpyxl==3.1.3
pytest-asyncio==0.24.0
